+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.1 (Unreleased)
 
 * r/heroku_app: Protect against panic [GH-11]
+* r/heroku_app: always read all config vars [GH-17]
+* r/heroku_app: Handle updating an app's stack [GH-16]
 
 ## 0.1.0 (June 20, 2017)